refactor(api): move campaign participation method to right repo
