+++ api/src/prescription/campaign-participation/infrastructure/repositories/campaign-participation-repository.js
@@ -4,15 +4,17 @@ import { knex } from '../../../../../db/knex-database-connection.js';
 import * as knowledgeElementRepository from '../../../../../lib/infrastructure/repositories/knowledge-element-repository.js';
 import * as knowledgeElementSnapshotRepository from '../../../../../lib/infrastructure/repositories/knowledge-element-snapshot-repository.js';
 import { Campaign } from '../../../../../src/shared/domain/models/Campaign.js';
+import { constants } from '../../../../shared/domain/constants.js';
 import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.js';
 import { NotFoundError } from '../../../../shared/domain/errors.js';
 import { Assessment } from '../../../../shared/domain/models/Assessment.js';
+import { CampaignParticipationStatuses, CampaignTypes } from '../../../shared/domain/constants.js';
 import { CampaignParticipation } from '../../domain/models/CampaignParticipation.js';
 import { AvailableCampaignParticipation } from '../../domain/read-models/AvailableCampaignParticipation.js';
 
-const { pick } = lodash;
+const { TO_SHARE } = CampaignParticipationStatuses;
 
-import { CampaignParticipationStatuses } from '../../../shared/domain/constants.js';
+const { pick } = lodash;
 
 const CAMPAIGN_PARTICIPATION_ATTRIBUTES = [
   'participantExternalId',
@@ -152,6 +154,48 @@ const findOneByCampaignIdAndUserId = async function ({ campaignId, userId }) {
   });
 };
 
+const hasAssessmentParticipations = async function (userId) {
+  const { count } = await knex('campaign-participations')
+    .count('campaign-participations.id')
+    .join('campaigns', 'campaigns.id', 'campaignId')
+    .whereNot('campaigns.organizationId', constants.AUTONOMOUS_COURSES_ORGANIZATION_ID)
+    .where('campaigns.type', '=', CampaignTypes.ASSESSMENT)
+    .andWhere({ userId })
+    .first();
+  return count > 0;
+};
+
+const getCodeOfLastParticipationToProfilesCollectionCampaignForUser = async function (userId) {
+  const result = await knex('campaign-participations')
+    .select('campaigns.code')
+    .join('campaigns', 'campaigns.id', 'campaignId')
+    .where({ userId })
+    .whereNull('campaign-participations.deletedAt')
+    .whereNull('archivedAt')
+    .andWhere({ status: TO_SHARE })
+    .andWhere({ 'campaigns.type': CampaignTypes.PROFILES_COLLECTION })
+    .orderBy('campaign-participations.createdAt', 'desc')
+    .first();
+  return result?.code || null;
+};
+
+const isRetrying = async function ({ campaignParticipationId }) {
+  const { id: campaignId, userId } = await knex('campaigns')
+    .select('campaigns.id', 'userId')
+    .join('campaign-participations', 'campaigns.id', 'campaignId')
+    .where({ 'campaign-participations.id': campaignParticipationId })
+    .first();
+
+  const campaignParticipations = await knex('campaign-participations')
+    .select('sharedAt', 'isImproved')
+    .where({ campaignId, userId });
+
+  return (
+    campaignParticipations.length > 1 &&
+    campaignParticipations.some((participation) => !participation.isImproved && !participation.sharedAt)
+  );
+};
+
 function _rowToResult(row) {
   return {
     id: row.id,
@@ -179,6 +223,9 @@ export {
   getAllCampaignParticipationsInCampaignForASameLearner,
   getByCampaignIds,
   getCampaignParticipationsForOrganizationLearner,
+  getCodeOfLastParticipationToProfilesCollectionCampaignForUser,
+  hasAssessmentParticipations,
+  isRetrying,
   remove,
   update,
   updateWithSnapshot,
