+++ api/tests/prescription/campaign-participation/integration/infrastructure/repositories/campaign-participation-repository_test.js
@@ -9,11 +9,12 @@ import {
   CampaignTypes,
 } from '../../../../../../src/prescription/shared/domain/constants.js';
 import { ApplicationTransaction } from '../../../../../../src/prescription/shared/infrastructure/ApplicationTransaction.js';
+import { constants } from '../../../../../../src/shared/domain/constants.js';
 import { NotFoundError } from '../../../../../../src/shared/domain/errors.js';
 import { Assessment } from '../../../../../../src/shared/domain/models/Assessment.js';
 import { catchErr, databaseBuilder, expect, knex, sinon } from '../../../../../test-helper.js';
 
-const { STARTED, SHARED } = CampaignParticipationStatuses;
+const { STARTED, SHARED, TO_SHARE } = CampaignParticipationStatuses;
 
 describe('Integration | Repository | Campaign Participation', function () {
   describe('#updateWithSnapshot', function () {
@@ -910,6 +911,376 @@ describe('Integration | Repository | Campaign Participation', function () {
 
       // then
       expect(response).to.equal(null);
+    });
+  });
+
+  describe('#hasAssessmentParticipations', function () {
+    let userId;
+
+    beforeEach(async function () {
+      sinon.stub(constants, 'AUTONOMOUS_COURSES_ORGANIZATION_ID').value(777);
+
+      userId = databaseBuilder.factory.buildUser().id;
+      await databaseBuilder.commit();
+    });
+
+    it('should return true if the user has participations to campaigns of type assement', async function () {
+      // given
+      const campaign = databaseBuilder.factory.buildCampaign({ type: CampaignTypes.ASSESSMENT });
+      databaseBuilder.factory.buildCampaignParticipation({
+        campaignId: campaign.id,
+        userId,
+      });
+      await databaseBuilder.commit();
+
+      // when
+      const result = await campaignParticipationRepository.hasAssessmentParticipations(userId);
+
+      // then
+      expect(result).to.equal(true);
+    });
+
+    it('should return false if the user does not have participations', async function () {
+      // given
+      const otherUser = databaseBuilder.factory.buildUser();
+      const campaign = databaseBuilder.factory.buildCampaign({ type: CampaignTypes.ASSESSMENT });
+      databaseBuilder.factory.buildCampaignParticipation({
+        campaignId: campaign.id,
+        userId: otherUser.id,
+      });
+      await databaseBuilder.commit();
+
+      // when
+      const result = await campaignParticipationRepository.hasAssessmentParticipations(userId);
+
+      // then
+      expect(result).to.equal(false);
+    });
+
+    it('should return false if the user does not have participations to campaigns of type assement', async function () {
+      // given
+      const campaign = databaseBuilder.factory.buildCampaign({ type: CampaignTypes.PROFILES_COLLECTION });
+      databaseBuilder.factory.buildCampaignParticipation({
+        campaignId: campaign.id,
+        userId,
+      });
+      await databaseBuilder.commit();
+
+      // when
+      const result = await campaignParticipationRepository.hasAssessmentParticipations(userId);
+
+      // then
+      expect(result).to.equal(false);
+    });
+
+    it('should return false if the user only have autonomouse-course participations', async function () {
+      // given
+      const organization = databaseBuilder.factory.buildOrganization({
+        id: constants.AUTONOMOUS_COURSES_ORGANIZATION_ID,
+      });
+      const campaign = databaseBuilder.factory.buildCampaign({
+        organizationId: organization.id,
+        type: CampaignTypes.ASSESSMENT,
+      });
+      databaseBuilder.factory.buildCampaignParticipation({
+        campaignId: campaign.id,
+        userId,
+      });
+      await databaseBuilder.commit();
+
+      // when
+      const result = await campaignParticipationRepository.hasAssessmentParticipations(userId);
+
+      // then
+      expect(result).to.equal(false);
+    });
+  });
+
+  describe('#getCodeOfLastParticipationToProfilesCollectionCampaignForUser', function () {
+    let userId;
+    const expectedCode = 'GOOD';
+
+    beforeEach(async function () {
+      userId = databaseBuilder.factory.buildUser().id;
+      await databaseBuilder.commit();
+    });
+
+    it('should return null if there is no participations', async function () {
+      // when
+      const code =
+        await campaignParticipationRepository.getCodeOfLastParticipationToProfilesCollectionCampaignForUser(userId);
+
+      // then
+      expect(code).to.equal(null);
+    });
+
+    it('should return null if there is no participations to share', async function () {
+      // given
+      const campaign = databaseBuilder.factory.buildCampaign({ type: 'PROFILES_COLLECTION' });
+      databaseBuilder.factory.buildCampaignParticipation({
+        campaignId: campaign.id,
+        status: SHARED,
+        userId,
+      });
+      await databaseBuilder.commit();
+
+      // when
+      const code =
+        await campaignParticipationRepository.getCodeOfLastParticipationToProfilesCollectionCampaignForUser(userId);
+
+      // then
+      expect(code).to.equal(null);
+    });
+
+    it('should return null if participations are deleted', async function () {
+      // given
+      const campaign = databaseBuilder.factory.buildCampaign({ type: 'PROFILES_COLLECTION' });
+      databaseBuilder.factory.buildCampaignParticipation({
+        campaignId: campaign.id,
+        status: TO_SHARE,
+        deletedAt: new Date(),
+        userId,
+      });
+      await databaseBuilder.commit();
+
+      // when
+      const code =
+        await campaignParticipationRepository.getCodeOfLastParticipationToProfilesCollectionCampaignForUser(userId);
+
+      // then
+      expect(code).to.equal(null);
+    });
+
+    it('should return null if there is no campaigns of type profiles collection', async function () {
+      const campaign = databaseBuilder.factory.buildCampaign({ type: 'ASSESSMENT' });
+      databaseBuilder.factory.buildCampaignParticipation({
+        campaignId: campaign.id,
+        status: TO_SHARE,
+        userId,
+      });
+      await databaseBuilder.commit();
+
+      // when
+      const code =
+        await campaignParticipationRepository.getCodeOfLastParticipationToProfilesCollectionCampaignForUser(userId);
+
+      // then
+      expect(code).to.equal(null);
+    });
+
+    it('should return null if there is no participations for the user', async function () {
+      const otherUser = databaseBuilder.factory.buildUser();
+      const campaign = databaseBuilder.factory.buildCampaign({ type: 'PROFILES_COLLECTION' });
+      databaseBuilder.factory.buildCampaignParticipation({
+        campaignId: campaign.id,
+        status: TO_SHARE,
+        userId: otherUser.id,
+      });
+      await databaseBuilder.commit();
+
+      // when
+      const code =
+        await campaignParticipationRepository.getCodeOfLastParticipationToProfilesCollectionCampaignForUser(userId);
+
+      // then
+      expect(code).to.equal(null);
+    });
+
+    it('should return null if campaign is archived', async function () {
+      // given
+      const campaign = databaseBuilder.factory.buildCampaign({ type: 'PROFILES_COLLECTION', archivedAt: new Date() });
+      databaseBuilder.factory.buildCampaignParticipation({
+        campaignId: campaign.id,
+        status: TO_SHARE,
+        userId,
+        createdAt: new Date(),
+      });
+      await databaseBuilder.commit();
+
+      // when
+      const code =
+        await campaignParticipationRepository.getCodeOfLastParticipationToProfilesCollectionCampaignForUser(userId);
+
+      // then
+      expect(code).to.equal(null);
+    });
+
+    it('should return code of the last participation to a campaign of type PROFILES_COLLECTION for the user', async function () {
+      // given
+      const campaign = databaseBuilder.factory.buildCampaign({ type: 'PROFILES_COLLECTION', code: expectedCode });
+      const otherCampaign = databaseBuilder.factory.buildCampaign({ type: 'PROFILES_COLLECTION', code: 'BAD' });
+      databaseBuilder.factory.buildCampaignParticipation({
+        campaignId: otherCampaign.id,
+        status: TO_SHARE,
+        createdAt: new Date(Date.parse('11/11/2011')),
+        userId,
+      });
+      databaseBuilder.factory.buildCampaignParticipation({
+        campaignId: campaign.id,
+        status: TO_SHARE,
+        userId,
+        createdAt: new Date(Date.parse('12/11/2011')),
+      });
+      await databaseBuilder.commit();
+
+      // when
+      const code =
+        await campaignParticipationRepository.getCodeOfLastParticipationToProfilesCollectionCampaignForUser(userId);
+
+      // then
+      expect(code).to.equal(expectedCode);
+    });
+  });
+
+  describe('#isRetrying', function () {
+    let campaignId;
+    let campaignParticipationId;
+    let userId;
+
+    beforeEach(async function () {
+      userId = databaseBuilder.factory.buildUser().id;
+      campaignId = databaseBuilder.factory.buildCampaign().id;
+      await databaseBuilder.commit();
+    });
+
+    context('When the user has just one participation shared', function () {
+      beforeEach(async function () {
+        campaignParticipationId = databaseBuilder.factory.buildCampaignParticipation({
+          campaignId,
+          userId,
+          isImproved: false,
+          sharedAt: new Date('2002-10-10'),
+        }).id;
+        await databaseBuilder.commit();
+      });
+
+      it('returns false', async function () {
+        const result = await campaignParticipationRepository.isRetrying({ userId, campaignParticipationId });
+        expect(result).to.be.false;
+      });
+    });
+
+    context('When the user has just one participation not shared', function () {
+      beforeEach(async function () {
+        campaignParticipationId = databaseBuilder.factory.buildCampaignParticipation({
+          campaignId,
+          userId,
+          isImproved: false,
+          sharedAt: null,
+        }).id;
+        await databaseBuilder.commit();
+      });
+
+      it('returns false', async function () {
+        const result = await campaignParticipationRepository.isRetrying({ userId, campaignParticipationId });
+        expect(result).to.be.false;
+      });
+    });
+
+    context('When the user has several participations but all shared', function () {
+      beforeEach(async function () {
+        databaseBuilder.factory.buildCampaignParticipation({
+          campaignId,
+          userId,
+          isImproved: true,
+          sharedAt: new Date('2002-10-10'),
+        });
+        campaignParticipationId = databaseBuilder.factory.buildCampaignParticipation({
+          campaignId,
+          userId,
+          isImproved: false,
+          sharedAt: new Date('2002-10-10'),
+        }).id;
+        await databaseBuilder.commit();
+      });
+
+      it('returns false', async function () {
+        const result = await campaignParticipationRepository.isRetrying({ userId, campaignParticipationId });
+        expect(result).to.be.false;
+      });
+    });
+
+    context('When the user has several participations but not in the same campaign', function () {
+      beforeEach(async function () {
+        const otherCampaignId = databaseBuilder.factory.buildCampaign().id;
+        databaseBuilder.factory.buildCampaignParticipation({
+          campaignId: otherCampaignId,
+          userId,
+          isImproved: true,
+          sharedAt: new Date('2002-10-10'),
+        });
+        databaseBuilder.factory.buildCampaignParticipation({
+          campaignId: otherCampaignId,
+          userId,
+          isImproved: false,
+          sharedAt: null,
+        });
+        campaignParticipationId = databaseBuilder.factory.buildCampaignParticipation({
+          campaignId,
+          userId,
+          isImproved: false,
+          sharedAt: null,
+        }).id;
+        await databaseBuilder.commit();
+      });
+
+      it('returns false', async function () {
+        const result = await campaignParticipationRepository.isRetrying({ userId, campaignParticipationId });
+        expect(result).to.be.false;
+      });
+    });
+
+    context('When there is several participations but not for the same user', function () {
+      beforeEach(async function () {
+        const otherUserId = databaseBuilder.factory.buildUser().id;
+        databaseBuilder.factory.buildCampaignParticipation({
+          campaignId,
+          userId: otherUserId,
+          isImproved: true,
+          sharedAt: new Date('2002-10-10'),
+        });
+        databaseBuilder.factory.buildCampaignParticipation({
+          campaignId,
+          userId: otherUserId,
+          isImproved: false,
+          sharedAt: null,
+        });
+        campaignParticipationId = databaseBuilder.factory.buildCampaignParticipation({
+          campaignId,
+          userId,
+          isImproved: false,
+          sharedAt: null,
+        }).id;
+        await databaseBuilder.commit();
+      });
+
+      it('returns false', async function () {
+        const result = await campaignParticipationRepository.isRetrying({ userId, campaignParticipationId });
+        expect(result).to.be.false;
+      });
+    });
+
+    context('When the user is retrying the campaign', function () {
+      beforeEach(async function () {
+        databaseBuilder.factory.buildCampaignParticipation({
+          campaignId,
+          userId,
+          isImproved: true,
+          sharedAt: new Date('2002-10-10'),
+        });
+        campaignParticipationId = databaseBuilder.factory.buildCampaignParticipation({
+          campaignId,
+          userId,
+          isImproved: false,
+          sharedAt: null,
+        }).id;
+        await databaseBuilder.commit();
+      });
+
+      it('returns true', async function () {
+        const result = await campaignParticipationRepository.isRetrying({ userId, campaignParticipationId });
+        expect(result).to.be.true;
+      });
     });
   });
 });
